Change nvgpu revision to resolve checkout failures
The current revision (21d928824dc7ca3dc17603a53b11edc6641ace2d) of nvgpu
submodule on main branch of nvidia-kernel-oot is orphaned and does not
belong to any branch or tag of the nvgpu repository. This causes issues
when checking out the submodule revision, for example when using repo,
or even simply cloning the nvgpu repository directly will also not
contain the 21d92882 ref. This doesn't seem to be happening when cloning
nvidia-kernel-oot directly using --recurse-submodules.

This updates the revision of nvgpu to the last parent commit starting
from the 21d928824dc7ca3dc17603a53b11edc6641ace2d ref that actually
belongs to a branch.